move some of the post query helper functions to utils/query-utils.js; allow user to specify an image url or upload an image for a post
